# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MotoMedic IMS API Documentation",
        "description": "API documentation for the MotoMedic Inventory Management System, generated with a modular structure.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/roles/{role}/permissions": {
            "post": {
                "tags": [
                    "RolePermissions"
                ],
                "summary": "Assign permissions to a role",
                "description": "Assigns permissions to a role",
                "operationId": "assignPermissions",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Permissions to assign",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RolePermissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WrappedRolePermissionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Error Response",
                "properties": {
                    "error": {
                        "description": "Error message.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "WrappedErrorResponse": {
                "title": "Wrapped Error Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "type": "object"
            },
...
            "RolePermissionRequest": {
                "title": "RolePermission Request",
                "required": [
                    "permissions"
                ],
                "properties": {
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "RolePermission": {
                "title": "RolePermission",
                "description": "RolePermission model",
                "properties": {
                    "id": {
                        "description": "Permission ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Permission name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "WrappedRolePermissionResponse": {
                "title": "Wrapped RolePermission Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Permissions assigned successfully"
                    },
                    "data": {
                        "properties": {
                            "role_name": {
                                "type": "string"
                            },
                            "permissions": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RolePermission"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Unauthorized. The token is invalid or expired.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "The specified resource was not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "UnprocessableEntity": {
                "description": "Unprocessable Entity",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/WrappedErrorResponse"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter 'Bearer' [space] and then your ACCESS TOKEN.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "refreshToken": {
                "type": "http",
                "description": "Enter 'Bearer' [space] and then your REFRESH TOKEN.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}